requirements.txt
wandb
ruamel.yaml
configmypy
tensorly
tensorly-torch
torch-harmonics
matplotlib
opt-einsum
h5py
zarr
fairscale
